+++ LE1_Stack/Makefile
@@ -1,0 +1,2 @@
+stackArray:
+	g++ Stack.cpp -o Stack
