+++ LE1_Stack/Stack.cpp
@@ -1,58 +1,81 @@
+#ifndef ABSTRACT_STACK_H
+#define ABSTRACT_STACK_H
+
+#include <stdexcept>
 #include <iostream>
-#include "Stack.h"
 
 using namespace std;
 
-class Stack : public AbstractStack<int>{
+// The AbstractStack class defies the interface for a stack.
+// Implement all the virtual methods in this class. 
+
+class AbstractStack {
+private:
+    // Your data here...
+
+public:
+    AbstractStack() {
+    }
+
+    ~AbstractStack() {
+    }
+
+    virtual bool isEmpty() = 0;
+    virtual int size() = 0;
+
+    // * Returns top element.
+    // * Throw an exception if the stack is empty.
+    virtual int top() = 0;
+
+    // * Removes and returns top element.
+    // * Throw an exception if the stack is empty.
+    virtual int pop() = 0;
+
+    // * Declared as pure virtual.
+    // * Overload this function in your child class!
+    virtual void push (int e) = 0;
+};
+
+#endif
+
+#ifndef STACK_ARRAY_H
+#define STACK_ARRAY_H
+
+class Stack : public AbstractStack{
 private:
     int* data;
-    int size;
-    int top;
+    int length;
+    int topIndex;
 
 public:
     Stack(){
-        size = 10;
-        data = new int[size];
-        top = -1;
     }
 
     ~Stack(){
-        delete[] data;
     }
 
     bool isEmpty(){
-        return top == -1;
+        return false;
     }
 
     int size(){
-        return size;
+        return 0;
     }
 
     int top(){
-        if(isEmpty()){
-            throw std::exception();
-        }
-        return data[top];
+        return 0;
     }
 
     int pop(){
-        if(isEmpty()){
-            throw std::exception();
-        }
-        return data[top--];
+        return 0;
     }
 
     void push(int e){
-        if(top == size - 1){
-            int* temp = new int[size * 2];
-            for(int i = 0; i < size; i++){
-                temp[i] = data[i];
-            }
-            delete[] data;
-            data = temp;
-            size *= 2;
-        }
-        data[++top] = e;
     }
 
 };
+#endif
+
+int main(){
+    cout<<"Hello World"<<endl;
+}
