+++ LE1_Stack/solution/test/testStackwithDoubleArray.cpp
@@ -1,4 +1,4 @@
-#include "../solution/Stack.h"
+#include "../Stack.cpp"
 #include <iostream>
 
 using namespace std;
