+++ LE1_Stack/solution/testStackwithDoubleArray
